requirements.txt
cloudscraper
validators
requests
pillow